+++ config-targets.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+export LANG=en_US.UTF-8
+export LC_ALL=en_US.UTF-8
+
 echo "###########################################"
 echo "#                                         #"
 echo "# Configuring Browsers Compatible Targets #"
@@ -16,6 +19,8 @@ if [ -f "$version_file" ]; then
 fi
 echo "-> Current Version v$VERSION"
 
+echo ""
+
 # BROWSERS Compabilities
 file_path="BROWSERS"
 
@@ -31,12 +36,20 @@ done < "$file_path"
 # Print the array elements to verify
 echo "-> Compatible Browsers: ${browsers[@]}"
 
-# Iterate over the array \xE2\x9C\x94
+echo ""
+
+# Iterate over the array ✓
 echo "-> Creting Targets"
 for browser in "${browsers[@]}"; do
+
+    echo ""
+
     echo "-> Creating target-$browser.sh"
     touch "target-$browser.sh"
     echo "sh target.sh $browser" > "target-$browser.sh"
-    echo " \xE2\x9C\x94 target-$browser.sh created."
+    echo " ✓ target-$browser.sh created."
 done
-echo "\n \xE2\x9C\x94 Targets ${browsers[@]} created!"
+
+echo ""
+
+echo " ✓ Targets ${browsers[@]} created!"